requirements.txt
azure-eventhub
numpy
azure-monitor
azure-identity 